django_AMZ/requirements.txt
apturl==0.5.2
asgiref==3.5.2
attrs==19.3.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloud-init==22.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.7.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.1
django-allauth==0.51.0
django-bootstrap3==22.1
django-filter==22.1
django-tables2==2.4.1
djongo==1.3.6
dnspython==2.2.1
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
future==0.18.2
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
mongoengine==0.24.2
monotonic==1.5
more-itertools==4.2.0
mysql-connector-python==8.0.30
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.20.1
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pykerberos==1.2.4
pymacaroons==0.13.0
pymongo==4.2.0
PyMySQL==1.0.2
PyNaCl==1.3.0
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-snappy==0.6.1
python3-openid==3.2.0
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sqlparse==0.2.4
ssh-import-id==5.10
systemd-python==234
ubuntu-advantage-tools==27.10
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
xkit==0.0.0
zipp==1.0.0
